mutt: huge changes
I redid the whole struture, so now I only keep a single, unified set of
maildirs. To match that changed setup, I also had to  redo a lot of the
mutt config. The end-goal here will be to be able to replace gmail
entirely by self-hosted email, which is why I will be starting to
archive all the mail in the maildir, and stop relying on git. I used a
shellscript (that's over in on sulami/spielwiese) to get all the mail
that has been deleted since I started my automatic mail repo. I addition
to that, I dumped my whole gmail archive into this archive, going back
to 2011. This will allow me to use a mail client or basic unix tools to
search all the non-spam mail that will be kept permanentely.