msmtp: default to Fastmail if I use a custom address
msmtp tries to select the mail account based on the from-address set,
but doesn't seem to allow regex. You can manually override the account
by using a parameter, but I don't have a good point in Emacs to do so.
If no account matches, msmtp just uses the account called "default",
so by renaming the Fastmail account to that, it will fall back to
using that one.