isync: don't expunge, and keep the dates
I did spend quite some time the other day trying to figure out why
emails seemed to be disappearing, or at least not syncing properly to
Fastmail.

Turns out there is a potential "edge case" in isync (as in, "I assume
that affects actually a lot of people"):

Archiving from the inbox means the email is moved from the inbox to
the archive. Isync syncs the inbox first, and notices the "deletion"
of the email (well, it's gone). So it tells the remote to delete it.
But then when it encounters the same email when syncing the archive,
it doesn't recreate the email, or Fastmail does something funky, not
sure yet, but it never shows up.

As a result I'm just going to stop expunging (deleting) emails, and
instead just move them to the trash directory. Fastmail cleans that
automatically after some amount of time, or I just do it by hand.

On another note, why is CopyArrivalDate not a default option?! Without
it, uploading an email means that it gets dated for the time of
upload. Not great if you're forcing re-sync, or sync up a new remote.