emacs: fix sulami/org-return
This previously only worked with plain lists (- foo) which were not
indented. The new solution now checks the type of the list as well as
for the possibility of a checkbox and indentation and takes all those
factors into account.

The new code also uses `evil-org-open-below` instead of
`org-meta-return`, because the former deals better with actually
inserting the list item scaffolding. Also it is more consistent
between coming from normal mode and coming from insert mode.

For future reference, this is a useful function to get a lot of info
for the list item at `point`:

```
(defun sulami/list-item-debug ()
  (interactive)
  (let ((struct (org-list-struct)))
    (message "point: %s\nbegin: %s\nend: %s\ntype: %s\nbullet: %s(%s)\nindent: %s\ncheckbox: %s(%s)"
             (point)
             (org-list-get-item-begin)
             (org-list-get-item-end (org-in-item-p)
                                    struct)
             (org-list-get-list-type (org-in-item-p)
                                     struct
                                     (org-list-prevs-alist struct))
             (org-list-get-bullet (org-in-item-p)
                                          struct)
             (length (org-list-get-bullet (org-in-item-p)
                                          struct))
             (org-list-get-ind (org-in-item-p)
                               struct)
             (org-list-get-checkbox (org-in-item-p)
                                    struct)
             (length (org-list-get-checkbox (org-in-item-p)
                                            struct)))))
```