emacs: find-file when switching projects
Counsel defaults to a combination of open buffers & files, which is
painfully slow to filter through (and also uses a different fuzzy
matcher). This treats every project switch as opening a new project.

Another compounding factor here is that buffer names don't include the
path, so they're not all that useful.