notmuch: prevent Gmail duplicate emails from mistagging
This was quite a debugging experience. This also started happening
some time over the holidays, not sure why exactly.

Gmail sends us mails in the inbox twice, once in the inbox, and then
as part of "All Mail". This makes sense so far, but notmuch treats the
same email in two places as one email (their message IDs and such are
the same).

Because I have pre- and post-hooks setup in notmuch to vaguely sync
directory structure to tags and tags to directory structure, this
caused some confusion. Because a new email was present in "All Mail",
it would lose its "inbox" tag. And because it was present in the
"inbox" directory but would not have an "inbox" tag, it would then get
moved to "All Mail".

The result was that new mails would get automatically archived within
a few runs (the sequencing of the hooks and runs matters here).

The resolution is actually quite simple:

Because notmuch allows several files to map to a single email entry,
we can identify the emails that are present in only one location and
archive only those (with tags). The tag->directory syncing can stay
blissfully unaware.