emacs: change squiggly underlines to straight ones
I find those easier on the eyes.

This is a bit ugly, because it does straight indexing of the attribute
list, because it's not actually an attribute list. It would be cleaner
if I would walk the list to look at the keys, but so far it works for
me™