emacs: Fix plain list cancellation at the end of buffers
This was working at the actual end of the buffer, as long as there was
no trailing newline (which adds 1 to point-max). So in this case, if
the end of our item, and point-max are equal, and point is one before
that end (due to the trailing newline), we assume we're cancelling the
list item.

Technically that last character could be something other than a
newline, so this will not work on

- x<EOF>
  ^ cursor before this x

but this is very much an edge case.