emacs: Fix org-mode links with spaces
Links in org-mode can often have spaces in their target, especially
when referring to a line in a file, where the line content is used as
the link target. This whitespace can break paragraph filling, e.g. via
org-fill-paragraph.

Org-mode links already have an escape system for brackets (with \),
and can optionally also be sort-of urlencoded (but apparently not
really). This means substituting a space with %20 fixes the text flow
issue without breaking the link itself.

This patch wraps the link escaping function (the one that normally
does just \) to also substitute spaces. This means that the normal
org-store-link; org-insert-link flow does the right thing.