emacs: path helpful-symbol to avoid inference errors
Helpful symbol is setup to propose the symbol at the cursor as the
symbol to show help for. While a great dwim idea, sometimes
helpful--symbol-at-point fails to parse whatever is under the cursor,
depending on the language in use, breaking the whole command.

This patch just wraps this suggestion code with a condition handler to
just not suggest anything if reading the symbol fails.