emacs: enable more projectile-based shells
I mainly use eshell, but shell and vterm are useful for the weirder
commands, such as nix-shell and anything taking over the whole
terminal. This just implements the same "switch to project shell if
exists, create otherwise" mechanism that I already have for eshell. A
bit of manual plumbing was required for vterm, which doesn't directly
integrate with projectile.