emacs: Get rid of org-mode buffer reverting
I used to have this because I was syncing my notes to my phone, so on
the sync back my computer would have stale buffers, which would mess
with the various refile-esque operations org can perform.

As I have since moved to org-roam, I don't really refile or archive
anymore, nor do I sync to my phone, so this is not needed anymore.

The actual reason to remove it is that org-roam does triggers this
quite a lot during regular operation, which leads to disruptive
slowdowns.